chore(cypress-ete): split reset password and sign in into individual tests
